feat(agent-mcp): let somebody else's agent connect, as themselves
The stdio server reads a session off one machine's disk. That is fine
for whoever owns the machine and useless for anybody else: handing it
out hands out a ledger. This is the other half.

`POST /api/mcp` on the web app takes a bearer token per request and
keeps nothing between them — no session, no key, nothing to leak. A
client with no token is refused with a `WWW-Authenticate` header naming
`/.well-known/oauth-protected-resource/api/mcp` (RFC 9728); it fetches
that, learns Supabase Auth is the authorization server, registers itself
and runs an authorization-code flow with PKCE. Nobody exchanges an API
key and Waves never holds one.

The token that comes back is an ordinary Supabase JWT, so every RLS
policy in this repo applies to a stranger's agent unchanged — which is
why this is a few hundred lines and not a second permission system. The
client is built with the anon key and that token; there is no
service-role key in the route, and there must never be one.

`buildWavesServer` moves out of the stdio entry into `tools.ts`, shared
by both transports: a tool that behaved differently depending on how the
request arrived would be a tool nobody could reason about. The exchange
itself goes through the SDK's web-standard transport, which speaks the
Request and Response a route handler already holds — the Node-shaped one
beside it would have meant hand-writing a fake `res`, and a hand-written
`res` silently misses whatever the SDK calls next. Stateless by
construction: one transport per request, closed after.

The internal imports lose their `.js` suffixes and the package moves to
bundler resolution. Nothing consumes it as plain node ESM any more —
tsx, vitest and Next's bundler — and NodeNext's mandatory suffixes
stopped the web app resolving the source at all. The `build`/`start`
scripts that pointed at a `dist/` no `noEmit` config ever produced are
gone with them.
