+++ apps/agent-mcp/package.json
@@ -4,13 +4,9 @@
   "private": true,
   "type": "module",
   "description": "An MCP server that lets an AI agent act inside Waves as a specific signed-in user — create groups, add expenses, and record settlements through the app's own authorized RPCs and edge functions, never raw SQL.",
-  "bin": {
-    "waves-agent-mcp": "dist/index.js"
-  },
   "scripts": {
-    "build": "tsc -p tsconfig.json",
     "dev": "tsx src/index.ts",
-    "start": "node dist/index.js",
+    "start": "tsx src/index.ts",
     "typecheck": "tsc -p tsconfig.json --noEmit",
     "test": "vitest run",
     "login": "tsx src/login.ts"
@@ -25,5 +21,11 @@
     "tsx": "^4.19.0",
     "typescript": "^5.6.0",
     "vitest": "^4.1.10"
+  },
+  "exports": {
+    ".": "./src/index.ts",
+    "./tools": "./src/tools.ts",
+    "./oauth": "./src/oauth.ts",
+    "./http": "./src/http.ts"
   }
 }
