+++ apps/web/package.json
@@ -19,7 +19,9 @@
     "@supabase/supabase-js": "^2.112.3",
     "next": "^16.3.0",
     "react": "19.2.8",
-    "react-dom": "19.2.8"
+    "react-dom": "19.2.8",
+    "@waves/agent-mcp": "workspace:*",
+    "@modelcontextprotocol/sdk": "^1.12.0"
   },
   "devDependencies": {
     "@playwright/test": "^1.62.1",
